AI/pyproject.toml
[tool.poetry]
name = "emilia-ai"
version = "0.1.0"
description = "Emilia - An AI-based Virtual Assistant for Emotional Wellbeing"
authors = ["Andre"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
langgraph = ">=0.0.25"
langchain = "^0.3.3"
langchain-openai = "^0.2.2"
langchain-community = "^0.3.13"
langchain-core = "^0.3.28"
python-dotenv = "^1.0.1"
jq = "^1.8.0"
langchain-google-genai = "^2.0.9"
pandas = "^2.2.3"
langchain-anthropic = "^0.2.3"
langchain-groq = "^0.2.4"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
pyyaml = "^6.0.2"
python-multipart = "0.0.17"
pydantic-settings = "^2.6.0"
python-json-logger = "^2.0.7"
pytz = "^2024.2"
rich = "^13.9.4"
langsmith = "^0.2.4"
langgraph-cli = { extras = ["inmem"], version = "^0.1.7" }
requests = "^2.32.3"
colorlog = "^6.9.0"
typing-extensions = "^4.12.2"
twilio = "^9.4.5"
streamlit = "^1.42.0"
jsonschema = "^4.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
